# models/socialmedia.go
package models

import (
	"github.com/asaskevich/govalidator"
	"gorm.io/gorm"
)

// Social media represents the model for a social media
type SocialMedia struct {
	GormModel
	UserID         uint
	Name           string `gorm:"not null" json:"name" form:"name" valid:"required~The name of social media is required"`
	SocialMediaUrl string `gorm:"not null" json:"social_media_url" form:"social_media_url" valid:"required~The social media URL is required"`
	User           *User
}

func (p *SocialMedia) BeforeCreate(tx *gorm.DB) (err error) {
	_, errCreate := govalidator.ValidateStruct(p)

	if errCreate != nil {
		err = errCreate
		return
	}
	err = nil
	return
}

func (p *SocialMedia) BeforeUpdate(tx *gorm.DB) (err error) {
	_, errUpdate := govalidator.ValidateStruct(p)

	if errUpdate != nil {
		err = errUpdate
		return
	}
	err = nil
	return
}